codes-from-ilia/defi Twap/requirements.txt
web3==6.15.1
python-dotenv==1.0.1
requests==2.31.0
eth-account==0.11.0
eth-typing==4.0.0
eth-utils==2.3.1
python-dateutil==2.8.2
pytz==2024.1
typing-extensions==4.9.0
cryptography==42.0.2
aiohttp==3.9.3
eth-abi==5.0.1
eth-hash==0.5.2
eth-keys==0.4.0
eth-rlp==0.3.0
hexbytes==0.3.1
jsonschema==4.21.1
lru-dict==1.2.0
protobuf==4.25.2
pydantic==2.6.1
websockets==12.0
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0 